Only trigger cleanup if the File is a directory

+++ core/src/main/java/net/microfalx/resource/FileResource.java
@@ -278,7 +278,7 @@ public class FileResource extends AbstractResource {
     @Override
     protected void afterEmpty() throws IOException {
         super.afterEmpty();
-        FileUtils.cleanDirectory(file);
+        if (file.isDirectory()) FileUtils.cleanDirectory(file);
     }
 
     private void createParent() throws IOException {
